requirements.txt
sentence-transformers
faiss-cpu
nltk
transformers
streamlit
python-docx
numpy
